+++ Cargo.toml
@@ -303,8 +303,8 @@ members = [
   "fuzz/fuzzer-libfuzzer",
   "gen-proxy-ffi",
   "proxy_ffi",
-  "proxy_tests",
   "proxy_test_raftstore_v2",
+  "proxy_tests",
   "raftstore-proxy",
   "raftstore-proxy-main",
   "tests",
