+++ mock-engine-store/Cargo.toml
@@ -42,6 +42,7 @@ log_wrappers = { workspace = true }
 pd_client = { workspace = true, default-features = false }
 protobuf = "2.8.0"
 proxy_server = { workspace = true }
+proxy_test_raftstore_v2 = { workspace = true }
 raft = { version = "0.7.0", default-features = false, features = ["protobuf-codec"] }
 raftstore = { workspace = true, default-features = false }
 raftstore-v2 = { workspace = true, default-features = false }
@@ -56,7 +57,6 @@ slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global
 tempfile = "3.0"
 test_pd_client = { workspace = true }
 test_raftstore = { workspace = true }
-proxy_test_raftstore_v2 = { workspace = true }
 test_util = { workspace = true, default-features = false }
 tikv = { workspace = true, default-features = false }
 tikv_util = { workspace = true, default-features = false }
