+++ proxy_tests/Cargo.toml
@@ -68,6 +68,7 @@ paste = "1.0"
 pd_client = { workspace = true }
 protobuf = { version = "2.8", features = ["bytes"] }
 proxy_server = { workspace = true, default-features = false }
+proxy_test_raftstore_v2 = { workspace = true }
 raft = { version = "0.7.0", default-features = false, features = ["protobuf-codec"] }
 raft_log_engine = { workspace = true }
 raftstore = { workspace = true }
