+++ proxy_tests/proxy/engine.rs
@@ -2,7 +2,7 @@
 
 use engine_traits::WriteBatchExt;
 
-use crate::utils::*;
+use crate::utils::v1::*;
 
 mod pagestorage {
     use super::*;
