+++ proxy_tests/proxy/mod.rs
@@ -1,3 +1,5 @@
+// Copyright 2023 TiKV Project Authors. Licensed under Apache-2.0.
+
 #![feature(box_patterns)]
 #![feature(test)]
 #![feature(custom_test_frameworks)]
@@ -13,7 +15,6 @@ mod config;
 mod engine;
 mod fast_add_peer;
 mod ffi;
-mod flashback;
 mod ingest;
 mod mock;
 mod normal;
@@ -21,5 +22,8 @@ mod region;
 mod replica_read;
 mod server_cluster_test;
 mod snapshot;
+mod store;
 mod utils;
+mod v1_specific;
+mod v2_specific;
 mod write;
