+++ proxy_tests/proxy/v1_specific/mod.rs
@@ -1,0 +1,4 @@
+// Copyright 2023 TiKV Project Authors. Licensed under Apache-2.0.
+
+mod flashback;
+mod region_ext;
